+++ src/main/java/org/truffleruby/core/exception/CoreExceptions.java
@@ -646,7 +646,11 @@ public class CoreExceptions {
     public DynamicObject nameErrorUndefinedMethod(String name, DynamicObject module, Node currentNode) {
         assert RubyGuards.isRubyModule(module);
         return nameError(
-                StringUtils.format("undefined method `%s' for %s", name, Layouts.MODULE.getFields(module).getName()),
+                StringUtils.format(
+                        "undefined method `%s' for %s `%s'",
+                        name,
+                        Layouts.CLASS.isClass(module) ? "class" : "module",
+                        Layouts.MODULE.getFields(module).getName()),
                 module,
                 name,
                 currentNode);